requirements.txt
docker==5.0.0
flask==2.0.1
flask-login==0.5.0
requests==2.22.0
six==1.16.0
ldap3==2.9.1
pyjwt==1.7.1
cryptography==3.4.8
pyyaml==5.3.1
flask-socketio==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
parameterized==0.8.1
apscheduler==3.8.1
